added legend and some other layers to play around with:
