+++ package.json
@@ -76,6 +76,7 @@
     "react-hot-loader": "^3.0.0-beta.6",
     "react-redux": "^5.0.3",
     "redux": "^3.6.0",
+    "redux-thunk": "^2.2.0",
     "request": "^2.81.0",
     "webpack": "^2.1.0-beta.6",
     "xhr": "^2.4.0",
