+++ src/actions/const.js
@@ -1,2 +1,17 @@
 /* eslint-disable import/prefer-default-export */
 /* Populated by react-webpack-redux:action */
+export const SET_CURRENT_SCALE = 'SET_CURRENT_SCALE';
+export const TOGGLE_LEGEND_NODE_VISIBLE = 'TOGGLE_LEGEND_NODE_VISIBLE';
+export const RESET_LEGEND_IS_FETCHING = 'RESET_LEGEND_IS_FETCHING';
+export const REQUEST_LEGEND_DATA = 'REQUEST_LEGEND_DATA';
+export const RECEIVE_LEGEND_DATA = 'RECEIVE_LEGEND_DATA';
+export const TOGGLE_LEGEND_NODE_EXPANDED = 'TOGGLE_LEGEND_NODE_EXPANDED';
+export const SET_LEGEND_DOM_DATA = 'SET_LEGEND_DOM_DATA';
+export const TOGGLE_SHOW_SETTINGS = 'TOGGLE_SHOW_SETTINGS';
+export const REVERSE_LAYER_ORDER = 'REVERSE_LAYER_ORDER';
+export const SHOW_LAYERS_NOT_VISIBLE_FOR_SCALE = 'SHOW_LAYERS_NOT_VISIBLE_FOR_SCALE';
+export const INIT_MAP_OPTIONS = 'INIT_MAP_OPTIONS';
+export const SET_INITIAL_LEGEND_MAPIMAGELAYER_DATA = 'SET_INITIAL_LEGEND_MAPIMAGELAYER_DATA';
+export const SET_INITIAL_LEGEND_GRAPHICSLAYER_DATA = 'SET_INITIAL_LEGEND_GRAPHICSLAYER_DATA';
+export const TOGGLE_LEGEND_EXPANDED = 'TOGGLE_LEGEND_EXPANDED';
+export const FETCH_LEGEND = 'FETCH_LEGEND';
