+++ src/actions/legend/toggleExpanded.js
@@ -1,0 +1,10 @@
+import { TOGGLE_LEGEND_EXPANDED } from './../const'
+
+function action(mapId, expanded) {
+  return {
+     type: TOGGLE_LEGEND_EXPANDED,
+     payload: { mapId, expanded }
+   }
+}
+
+module.exports = action;
