+++ src/components/legend/legend.cssmodule.scss
@@ -1,0 +1,3 @@
+.legend-component {
+  border: 1px dashed #f00;
+}
