+++ src/containers/App.js
@@ -10,13 +10,26 @@ import React, {
 } from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import '../actions/';
-import Main from '../components/App';
+import {
+  set_initial_legend,
+  setInitialLegend,
+  reverseLayerOrder,
+  showLayersNotVisibleForScale,
+  toggleExpanded,
+  toggleNodeExpanded,
+  toggleNodeVisible,
+  toggleShowSettings,
+  fetchLegend
+} from '../actions/';
+import AppComponent from '../components/App';
 /* Populated by react-webpack-redux:reducer */
 class App extends Component {
   render() {
-    const {actions} = this.props;
-    return <Main actions={actions}/>;
+    const {actions, mapLegend, initLegend} = this.props;
+    return <AppComponent
+      actions={actions}
+      initLegend={initLegend}
+      mapLegend={mapLegend}/>;
   }
 }
 /* Populated by react-webpack-redux:reducer
@@ -25,19 +38,38 @@ class App extends Component {
  *       adjust it here.
  */
 App.propTypes = {
-  actions: PropTypes.shape({})
+  actions: PropTypes.shape({
+    setInitialLegend: PropTypes.func.isRequired,
+    reverseLayerOrder: PropTypes.func.isRequired,
+    showLayersNotVisibleForScale: PropTypes.func.isRequired,
+    toggleExpanded: PropTypes.func.isRequired,
+    toggleNodeExpanded: PropTypes.func.isRequired,
+    toggleNodeVisible: PropTypes.func.isRequired,
+    toggleShowSettings: PropTypes.func.isRequired,
+    fetchLegend: PropTypes.func.isRequired
+  })
 };
 function mapStateToProps(state) {
   // eslint-disable-line no-unused-vars
   /* Populated by react-webpack-redux:reducer */
-  const props = {
-  };
+  const props = { mapLegend: state.mapLegend };
   return props;
 }
 function mapDispatchToProps(dispatch) {
   /* Populated by react-webpack-redux:action */
-  const actions = {};
-  const actionMap = { actions: bindActionCreators(actions, dispatch) };
+  const actions = {
+    setInitialLegend,
+    reverseLayerOrder,
+    showLayersNotVisibleForScale,
+    toggleExpanded,
+    toggleNodeExpanded,
+    toggleNodeVisible,
+    toggleShowSettings,
+    fetchLegend
+  };
+  const actionMap = {
+    actions: bindActionCreators(actions, dispatch)
+  };
   return actionMap;
 }
 export default connect(mapStateToProps, mapDispatchToProps)(App);
